Backed out changeset 98c7277a8fb5 (bug 1166350) for xpcshell failures.

+++ dom/push/PushService.jsm
@@ -116,26 +116,6 @@ this.PushService = {
     }
   },
 
-  _makePendingKey: function(aPageRecord) {
-    return aPageRecord.scope + "|" + aPageRecord.originAttributes;
-  },
-
-  _lookupOrPutPendingRequest: function(aPageRecord) {
-    let key = this._makePendingKey(aPageRecord);
-    if (this._pendingRegisterRequest[key]) {
-      return this._pendingRegisterRequest[key];
-    }
-
-    return this._pendingRegisterRequest[key] = this._registerWithServer(aPageRecord);
-  },
-
-  _deletePendingRequest: function(aPageRecord) {
-    let key = this._makePendingKey(aPageRecord);
-    if (this._pendingRegisterRequest[key]) {
-      delete this._pendingRegisterRequest[key];
-    }
-  },
-
   _setState: function(aNewState) {
     debug("new state: " + aNewState + " old state: " + this._state);
 
@@ -253,41 +233,24 @@ this.PushService = {
           return;
         }
 
-        var originAttributes =
-          ChromeUtils.originAttributesToSuffix({ appId: data.appId,
-                                                 inBrowser: data.browserOnly });
-        this._db.getAllByOriginAttributes(originAttributes)
-          .then(records => {
-            records.forEach(record => {
-              this._db.delete(this._service.getKeyFromRecord(record))
-                .then(_ => {
-                  // courtesy, but don't establish a connection
-                  // just for it
-                  if (this._ws) {
-                    debug("Had a connection, so telling the server");
-                    this._sendRequest("unregister", {channelID: records.channelID})
-                        .catch(function(e) {
-                          debug("Unregister errored " + e);
-                        });
-                  }
-                }, err => {
-                  debug("webapps-clear-data: " + scope +
-                        " Could not delete entry " + records.channelID);
+        // TODO 1149274.  We should support site permissions as well as a way to
+        // go from manifest url to 'all scopes registered for push in this app'
+        let appsService = Cc["@mozilla.org/AppsService;1"]
+                            .getService(Ci.nsIAppsService);
+        let scope = appsService.getScopeByLocalId(data.appId);
+        if (!scope) {
+          debug("webapps-clear-data: No scope found for " + data.appId);
+          return;
+        }
 
-                  // courtesy, but don't establish a connection
-                  // just for it
-                  if (this._ws) {
-                    debug("Had a connection, so telling the server");
-                    this._sendRequest("unregister", {channelID: records.channelID})
-                        .catch(function(e) {
-                          debug("Unregister errored " + e);
-                        });
-                  }
-                  throw "Database error";
-                });
-            });
-          }, _ => {
-            debug("webapps-clear-data: Error in getAllByOriginAttributes(" + originAttributes + ")");
+        this._db.getByScope(scope)
+          .then(record =>
+            Promise.all([
+              this._db.delete(this._service.getKeyFromRecord(record)),
+              this._sendRequest("unregister", record)
+            ])
+          ).catch(_ => {
+            debug("webapps-clear-data: Error in getByScope(" + scope + ")");
           });
 
         break;
@@ -660,9 +623,13 @@ this.PushService = {
     // records are objects describing the registration as stored in IndexedDB.
     return this._db.getAllKeyIDs()
       .then(records => {
+        let scopes = new Set();
+        for (let record of records) {
+          scopes.add(record.scope);
+        }
         let globalMM = Cc['@mozilla.org/globalmessagemanager;1']
                          .getService(Ci.nsIMessageListenerManager);
-        for (let record of records) {
+        for (let scope of scopes) {
           // Notify XPCOM observers.
           Services.obs.notifyObservers(
             null,
@@ -671,8 +638,8 @@ this.PushService = {
           );
 
           let data = {
-            originAttributes: record.originAttributes,
-            scope: record.scope
+            originAttributes: {}, // TODO bug 1166350
+            scope: scope
           };
 
           globalMM.broadcastAsyncMessage('pushsubscriptionchange', data);
@@ -692,7 +659,7 @@ this.PushService = {
         );
 
         let data = {
-          originAttributes: record.originAttributes,
+          originAttributes: {}, // TODO bug 1166350
           scope: record.scope
         };
 
@@ -714,7 +681,7 @@ this.PushService = {
           );
 
           let data = {
-            originAttributes: record.originAttributes,
+            originAttributes: {}, // TODO bug 1166350
             scope: record.scope
           };
 
@@ -728,8 +695,7 @@ this.PushService = {
   },
 
   _notifyApp: function(aPushRecord, message) {
-    if (!aPushRecord || !aPushRecord.scope ||
-        aPushRecord.originAttributes === undefined) {
+    if (!aPushRecord || !aPushRecord.scope) {
       debug("notifyApp() something is undefined.  Dropping notification: " +
         JSON.stringify(aPushRecord) );
       return;
@@ -762,7 +728,7 @@ this.PushService = {
     // TODO data.
     let data = {
       payload: message,
-      originAttributes: aPushRecord.originAttributes,
+      originAttributes: {}, // TODO bug 1166350
       scope: aPushRecord.scope
     };
 
@@ -779,7 +745,7 @@ this.PushService = {
     return this._db.getAllKeyIDs();
   },
 
-  _sendRequest: function(action, aRecord) {
+  _sendRequest(action, aRecord) {
     if (this._state == PUSH_SERVICE_CONNECTION_DISABLE) {
       return Promise.reject({state: 0, error: "Service not active"});
     } else if (this._state == PUSH_SERVICE_ACTIVE_OFFLINE) {
@@ -793,35 +759,37 @@ this.PushService = {
    * navigator.push, identifying the sending page and other fields.
    */
   _registerWithServer: function(aPageRecord) {
-    debug("registerWithServer()" + JSON.stringify(aPageRecord));
+    debug("registerWithServer()");
 
     return this._sendRequest("register", aPageRecord)
       .then(pushRecord => this._onRegisterSuccess(pushRecord),
             err => this._onRegisterError(err))
       .then(pushRecord => {
-        this._deletePendingRequest(aPageRecord);
+        if (this._pendingRegisterRequest[aPageRecord.scope]) {
+          delete this._pendingRegisterRequest[aPageRecord.scope];
+        }
         return pushRecord;
       }, err => {
-        this._deletePendingRequest(aPageRecord);
+        if (this._pendingRegisterRequest[aPageRecord.scope]) {
+          delete this._pendingRegisterRequest[aPageRecord.scope];
+        }
         throw err;
      });
   },
 
   _register: function(aPageRecord) {
-    debug("_register()");
-    if (!aPageRecord.scope || aPageRecord.originAttributes === undefined) {
-      return Promise.reject({state: 0, error: "NotFoundError"});
-    }
-
     return this._checkActivated()
-      .then(_ => this._db.getByIdentifiers(aPageRecord))
+      .then(_ => this._db.getByScope(aPageRecord.scope))
       .then(pushRecord => {
         if (pushRecord === undefined) {
-          return this._lookupOrPutPendingRequest(aPageRecord);
+          if (this._pendingRegisterRequest[aPageRecord.scope]) {
+            return this._pendingRegisterRequest[aPageRecord.scope];
+          }
+          return this._pendingRegisterRequest[aPageRecord.scope] = this._registerWithServer(aPageRecord);
         }
         return pushRecord;
       }, error => {
-        debug("getByIdentifiers failed");
+        debug("getByScope failed");
         throw error;
       });
   },
@@ -865,39 +833,10 @@ this.PushService = {
       return;
     }
 
-    if (!aMessage.target.assertPermission("push")) {
-      debug("Got message from a child process that does not have 'push' permission.");
-      return null;
-    }
-
     let mm = aMessage.target.QueryInterface(Ci.nsIMessageSender);
-    let pageRecord = aMessage.data;
+    let json = aMessage.data;
 
-    let principal = aMessage.principal;
-    if (!principal) {
-      debug("No principal passed!");
-      let message = {
-        requestID: aPageRecord.requestID,
-        error: "SecurityError"
-      };
-      mm.sendAsyncMessage("PushService:Register:KO", message);
-      return;
-    }
-
-    pageRecord.originAttributes =
-      ChromeUtils.originAttributesToSuffix(principal.originAttributes);
-
-    if (!pageRecord.scope || pageRecord.originAttributes === undefined) {
-      debug("Incorrect identifier values set! " + JSON.stringify(pageRecord));
-      let message = {
-        requestID: aPageRecord.requestID,
-        error: "SecurityError"
-      };
-      mm.sendAsyncMessage("PushService:Register:KO", message);
-      return;
-    }
-
-    this[aMessage.name.slice("Push:".length).toLowerCase()](pageRecord, mm);
+    this[aMessage.name.slice("Push:".length).toLowerCase()](json, mm);
   },
 
   register: function(aPageRecord, aMessageManager) {
@@ -944,12 +883,13 @@ this.PushService = {
    */
   _unregister: function(aPageRecord) {
     debug("_unregister()");
-    if (!aPageRecord.scope || aPageRecord.originAttributes === undefined) {
+
+    if (!aPageRecord.scope) {
       return Promise.reject({state: 0, error: "NotFoundError"});
     }
 
     return this._checkActivated()
-      .then(_ => this._db.getByIdentifiers(aPageRecord))
+      .then(_ => this._db.getByScope(aPageRecord.scope))
       .then(record => {
         // If the endpoint didn't exist, let's just fail.
         if (record === undefined) {
@@ -992,12 +932,12 @@ this.PushService = {
    */
   _registration: function(aPageRecord) {
     debug("_registration()");
-    if (!aPageRecord.scope || aPageRecord.originAttributes === undefined) {
-      return Promise.reject({state: 0, error: "NotFoundError"});
+    if (!aPageRecord.scope) {
+      return Promise.reject({state: 0, error: "Database error"});
     }
 
     return this._checkActivated()
-      .then(_ => this._db.getByIdentifiers(aPageRecord))
+      .then(_ => this._db.getByScope(aPageRecord.scope))
       .then(pushRecord => {
         if (!pushRecord) {
           return null;
